requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
autopep8==2.0.2
bcrypt==4.0.1
boto3==1.28.1
botocore==1.31.1
cffi==1.15.1
click==8.1.4
colorama==0.4.6
cryptography==41.0.3
dependency-injector==4.41.0
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-sessions==0.3.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
jmespath==1.0.1
passlib==1.7.4
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.0.2
pydantic-settings==2.0.3
pydantic_core==2.1.2
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
redis==4.6.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
